+++ packages/atomic/src/components/facets/atomic-category-facet/atomic-category-facet.tsx
@@ -1,4 +1,4 @@
-import {Component, h, Prop, State, Host} from '@stencil/core';
+import {Component, h, Prop, State, Host, Element} from '@stencil/core';
 import {
   CategoryFacetState,
   CategoryFacet,
@@ -31,6 +31,7 @@ import {
   FacetSearchStrings,
 } from '../facet-search/facet-search';
 import {FacetPlaceholder} from '../atomic-facet-placeholder/atomic-facet-placeholder';
+import {registerFacetToStore} from '../../../utils/store';
 
 const SEPARATOR = '/';
 const ELLIPSIS = '...';
@@ -75,6 +76,7 @@ export class AtomicCategoryFacet
   @InitializeBindings() public bindings!: Bindings;
   public facet!: CategoryFacet;
   public searchStatus!: SearchStatus;
+  @Element() private host!: HTMLElement;
 
   @BindStateToController('facet')
   @State()
@@ -164,9 +166,11 @@ export class AtomicCategoryFacet
       this.facetSearch = new FacetSearch(this);
     }
     this.facetId = this.facet.state.facetId;
-    this.bindings.store.state.categoryFacets[this.facetId] = {
+    registerFacetToStore(this.bindings.store, 'categoryFacets', {
       label: this.label,
-    };
+      facetId: this.facetId!,
+      element: this.host,
+    });
   }
 
   private get parents() {
@@ -188,7 +192,7 @@ export class AtomicCategoryFacet
         class="value-button"
         onClick={() => this.facet.toggleSelect(parent)}
       >
-        <div innerHTML={LeftArrow} class="facet-arrow mr-1.5" />
+        <atomic-icon icon={LeftArrow} class="facet-arrow mr-1.5"></atomic-icon>
         <span class="truncate">{parent.value}</span>
       </button>
     );
@@ -226,7 +230,10 @@ export class AtomicCategoryFacet
           <span part="value-count" class="value-count">
             {item.numberOfResults.toLocaleString(this.bindings.i18n.language)}
           </span>
-          <div innerHTML={RightArrow} class="facet-arrow ml-1.5" />
+          <atomic-icon
+            icon={RightArrow}
+            class="facet-arrow ml-1.5"
+          ></atomic-icon>
         </button>
       </li>
     );
@@ -243,7 +250,7 @@ export class AtomicCategoryFacet
         onClick={() => this.facet.deselectAll()}
         class="value-button"
       >
-        <div innerHTML={LeftArrow} class="facet-arrow mr-1.5" />
+        <atomic-icon icon={LeftArrow} class="facet-arrow mr-1.5"></atomic-icon>
         {this.strings.allCategories()}
       </button>
     );